requirements.txt
pydglab-ws
qrcode
python-osc
colorlog
pillow
ruamel.yaml
psutil
PySide6
qasync
websockets
requests
types-requests
mypy>=1.17.0
watchdog
pyinstaller